requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-generator==1.10
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.3
beautifulsoup4==4.11.2
blinker==1.8.2
botocore==1.31.37
cached-property==1.5.2
cachetools==5.3.1
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
contourpy==1.1.1
cryptography==42.0.7
cycler==0.12.0
dataclasses-json==0.6.6
DateTime==5.1
decorator==5.1.1
distro==1.8.0
dnspython==2.6.1
docarray==0.40.0
docutils==0.16
dominate==2.9.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.40
gTTS==2.5.1
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
idna==3.7
inflection==0.5.1
ipython
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jose==1.0.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
kiwisolver==1.4.5
langchain==0.2.0
langchain-openai==0.1.7
langchain-text-splitters==0.2.0
langserve==0.2.0
langsmith==0.1.59
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.8.0
matplotlib-inline==0.1.7
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.30.1
orjson==3.10.3
outcome==1.2.0
packaging==23.2
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
Pillow==10.0.1
pipdeptree==2.20.0
prompt-toolkit==3.0.43
proto-plus==1.22.3
protobuf==4.24.4
psycopg-binary==3.1.12
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.22
pydantic>=2.0
pydantic-settings==2.0.3
Pygments==2.18.0
pyparsing==3.1.1
pyproject-toml==0.0.10
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==2.0.0
rich==13.7.1
rsa==4.7.2
selenium==4.12.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.4
SQLAlchemy==2.0.30
sse-starlette==2.1.0
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0 
tenacity==8.3.0
tiktoken==0.7.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.22.0
trio-websocket==0.9.2
tweepy==4.14.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.10.0
urllib3==1.26.15
uvicorn==0.29.0
uvloop==0.19.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
yarl==1.9.4
zope.interface==6.0
